[rrc] move the driver into the library as reussir_compiler::driver
Mechanical: src/main.rs becomes a shim over driver::main(), the 1900
lines move to src/driver.rs unchanged (imports rebased onto crate::).
The driver gains a real module home so the next commits can split it —
and its pieces become unit-testable library code instead of bin-only
items.
