+++ crates/reussir-compiler/src/lib.rs
@@ -11,6 +11,7 @@
 //! the polymorphic-FFI gather), run the lowering pipeline, finalize, then hand
 //! the [`Finalized`] module to [`emit_to_file`].
 
+pub mod driver;
 pub mod package;
 
 use std::ffi::{CString, c_char};
